Move build-corpus into a cargo script (#51)
The build corpus package would be better served as a script. However it
was written as a Rust crate due to its dependency on `sleigh-config`.
However a Cargo script is more suitable for this task. It is currently
available in nightly:
https://github.com/rust-lang/cargo/issues/12207#issuecomment-3412997290.
The requirement for nightly isn't new since it is already required for
address sanitizer.

Now that the fuzzer is the only package in the workspace, this enables
removal of the Cargo workspace. Moved the fuzzer to the root of the
repository to simplify the organization.